--- PROJETOS/20180712-useacabeca-c/blasteroids/blast.h
@@ -1,13 +1,0 @@
-#ifndef BLASTEROIDS_BLAST
-#define BLASTEROIDS_BLAST
-
-typedef struct Blast {
-    float sx;
-    float sy;
-    float heading;
-    float speed;
-    bool isOn;
-    ALLEGRO_COLOR color;
-} Blast;
-
-#endif
