--- PROJETOS/20180712-useacabeca-c/blasteroids/draw_utils.h
@@ -1,8 +1,0 @@
-#ifndef BLASTEROIDS_DRAW_UTILS
-#define BLASTEROIDS_DRAW_UTILS
-
-float deg2rad(float deg);
-
-void blasteroids_get_delta(float *deltax, float *deltay, float speed, float degrees);
-
-#endif
