+++ PROJETOS/20180712-useacabeca-c/blasteroids/include/blasteroids.h
@@ -1,0 +1,17 @@
+#include <allegro5/allegro.h>
+#include <allegro5/allegro_primitives.h>
+#include <stdbool.h>
+
+#include <blasteroids/blasteroids_types.h>
+
+#define DISPLAY_ALTURA 600
+#define DISPLAY_LARGURA 600
+
+#include <blasteroids/utils.h>
+
+#include <blasteroids/spaceship.h>
+#include <blasteroids/asteroid.h>
+#include <blasteroids/blast.h>
+#include <blasteroids/main.h>
+#include <blasteroids/event.h>
+
