+++ PROJETOS/20180712-useacabeca-c/blasteroids/include/blasteroids/main.h
@@ -1,0 +1,14 @@
+#ifndef BLASTEROIDS_MAIN
+#define BLASTEROIDS_MAIN
+#include <stdbool.h>
+#include <blasteroids/blasteroids_types.h>
+
+bool is_collision(GameContext *ctx);
+
+void update_states();
+
+void handle_shutdown();
+
+void stop(int sig);
+
+#endif
