+++ PROJETOS/20180712-useacabeca-c/blasteroids/include/blasteroids/utils.h
@@ -1,0 +1,3 @@
+#include <blasteroids/utils/draw.h>
+#include <blasteroids/utils/log.h>
+#include <blasteroids/utils/signal.h>
