+++ PROJETOS/20180712-useacabeca-c/blasteroids/include/blasteroids/utils/draw.h
@@ -1,0 +1,10 @@
+#ifndef BLASTEROIDS_UTILS_DRAW
+#define BLASTEROIDS_UTILS_DRAW
+
+float deg2rad(float deg);
+
+void blasteroids_get_delta(float *deltax, float *deltay, float speed, float degrees);
+
+float get_distance(float ax, float ay, float bx, float by);
+
+#endif
