+++ PROJETOS/20180712-useacabeca-c/blasteroids/include/blasteroids/utils/signal.h
@@ -1,0 +1,6 @@
+#ifndef BLASTEROIDS_UTILS_SIGNAL
+#define BLASTEROIDS_UTILS_SIGNAL
+
+int catch_signal(int sig, void(*handler)(int));
+
+#endif
