--- PROJETOS/20180712-useacabeca-c/blasteroids/signal_utils.h
@@ -1,1 +1,0 @@
-int catch_signal(int sig, void(*handler)(int));
