+++ PROJETOS/20180712-useacabeca-c/blasteroids/src/utils/draw.c
@@ -1,6 +1,5 @@
-#include "log_utils.h"
-#include "draw_utils.h"
 #include <math.h>
+#include <blasteroids/utils/draw.h>
 
 float deg2rad(float deg) {
     return 0.0174532925 * deg;
@@ -9,4 +8,12 @@ float deg2rad(float deg) {
 void blasteroids_get_delta(float *deltax, float *deltay, float speed, float degrees) {
     *deltax = speed * sin((double)deg2rad(degrees));
     *deltay = speed * cos((double)deg2rad(degrees)) * -1;
+}
+
+
+float get_distance(float ax, float ay, float bx, float by) {
+    float x, y;
+    x = ax - bx; // Não temos distancia negativa
+    y = ay - by;
+    return sqrtf(x*x + y*y); // Vi va pitágoras :p
 }
