+++ PROJETOS/20180712-useacabeca-c/blasteroids/src/utils/log.c
